+++ src/Bem/Utl/Intr.hs
@@ -36,3 +36,21 @@ type FromElem (b :: (Type -> Type) -> Type -> Type) r
 type FromFullBlk (b :: (Type -> Type) -> Type -> Type) r
          =
          forall e m . (Show (b e m), Show m) => b e m -> [m] -> r
+
+-- | from a block and an element with their modifiers into something
+type FromBlkElem r = forall b . FromFullBlk b (FromFullElem b r)
+
+{- |
+from a block with its modifiers
+and an element without its modifiers into something
+-}
+type FromBlkNoModsElem r = forall b . FromFullBlk b (FromElem b r)
+
+{- |
+from a block without its modifiers
+and an element with its modifiers into something
+-}
+type FromNoModsBlkElem r = forall b . FromBlk b (FromFullElem b r)
+
+-- | from a block and an element without their modifiers into something
+type FromNoModsBlkNoModsElem r = forall b . FromBlk b (FromElem b r)