+++ src/Bem/Cls/Gen/Gen.hs
@@ -10,7 +10,7 @@ module Bem.Cls.Gen.Gen where
 
 import qualified Bem.Cls.Gen.Intr as IntrGen
 import Bem.Utl.Intr
-import Bem.Cfg.Cfg
+import qualified Bem.Cfg.Cfg as Cfg
 
 import Control.Monad.Reader
 
@@ -21,11 +21,11 @@ genBlkElem blk blkMods prntBlk elem' elemMods
     =
     runReader (IntrGen.genBlk blk blkMods prntBlk elem' elemMods) defCfg
   where
-    defCfg = Cfg { _elemSep = "__"
-                 , _modSep = "_"
-                 , _partSep = "-"
-                 , _partsAreCptled = False
-                 }
+    defCfg = Cfg.Cfg { Cfg._elemSep = "__"
+                     , Cfg._modSep = "_"
+                     , Cfg._partSep = "-"
+                     , Cfg._partsAreCptled = False
+                     }
 
 -- | Generate a class of a element along with its modifiers.
 genElem :: FromFullElem b Class
@@ -33,8 +33,8 @@ genElem prntBlk elem' elemMods
     =
     runReader (IntrGen.genElem prntBlk elem' elemMods) defCfg
   where
-    defCfg = Cfg { _elemSep = "__"
-                 , _modSep = "_"
-                 , _partSep = "-"
-                 , _partsAreCptled = False
-                 }
+    defCfg = Cfg.Cfg { Cfg._elemSep = "__"
+                     , Cfg._modSep = "_"
+                     , Cfg._partSep = "-"
+                     , Cfg._partsAreCptled = False
+                     }